@@ -25,63 +25,62 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.1.3",
-    "@angular/cdk": "^7.1.1",
-    "@angular/common": "7.1.3",
-    "@angular/compiler": "7.1.3",
-    "@angular/core": "7.1.3",
-    "@angular/forms": "7.1.3",
-    "@angular/material": "^7.1.1",
-    "@angular/platform-browser": "7.1.3",
-    "@angular/platform-browser-dynamic": "7.1.3",
-    "@angular/platform-server": "7.1.3",
-    "@angular/router": "7.1.3",
-    "@angular/service-worker": "7.1.3",
-    "@covalent/markdown": "^1.0.1",
-    "@covalent/text-editor": "^1.0.0",
+    "@angular/animations": "8.1.0",
+    "@angular/cdk": "^8.0.2",
+    "@angular/common": "8.1.0",
+    "@angular/compiler": "8.1.0",
+    "@angular/core": "8.1.0",
+    "@angular/forms": "8.1.0",
+    "@angular/material": "^8.0.2",
+    "@angular/platform-browser": "8.1.0",
+    "@angular/platform-browser-dynamic": "8.1.0",
+    "@angular/platform-server": "8.1.0",
+    "@angular/router": "8.1.0",
+    "@angular/service-worker": "8.1.0",
+    "@covalent/markdown": "^2.1.0",
+    "@covalent/text-editor": "^2.1.0",
     "@types/cropperjs": "^1.1.2",
     "chart.js": "^2.7.3",
-    "core-js": "^2.5.1",
     "cropperjs": "^1.3.4",
     "jszip": "^3.1.5",
     "material-design-icons": "^3.0.1",
-    "material-icons": "^0.1.0",
+    "material-icons": "^0.3.1",
     "mime": "^2.4.3",
     "ngx-img": "^10.15.0",
     "pouchdb": "^7.0.0",
     "pouchdb-authentication": "^1.1.3",
     "pouchdb-find": "^7.0.0",
-    "rxjs": "6.4.0",
+    "rxjs": "6.5.2",
     "showdown": "1.6.4",
-    "zone.js": "~0.8.26"
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.6",
-    "@angular/cli": "^7.1.3",
-    "@angular/compiler-cli": "7.1.3",
-    "@angular/language-service": "7.1.3",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
+    "@angular-devkit/build-angular": "~0.801.2",
+    "@angular/cli": "^8.1.0",
+    "@angular/compiler-cli": "8.1.0",
+    "@angular/language-service": "8.1.0",
+    "@types/jasmine": "~3.3.13",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~12.0.12",
     "@types/pouchdb": "^6.3.2",
     "@types/pouchdb-find": "^6.3.3",
-    "codelyzer": "~3.2.0",
+    "codelyzer": "^5.0.1",
     "htmlhint-ng2": "0.0.13",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "node-sass": "^4.5.3",
-    "protractor": "~5.1.2",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "~5.4.2",
     "request-promise": "^4.2.1",
-    "sass-lint": "^1.12.1",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "tslint-eslint-rules": "^4.1.1",
-    "typescript": "3.1.6"
+    "sass-lint": "^1.13.1",
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "tslint-eslint-rules": "^5.4.0",
+    "typescript": "3.4.5"
   }
 }
