@@ -18,7 +18,7 @@ export class ResourcesMenuComponent {
 
   @Input() resources: any[] = [];
   @Input() color = 'accent';
-  @ViewChild(MatMenuTrigger) resourceButton: MatMenuTrigger;
+  @ViewChild(MatMenuTrigger, { static: false }) resourceButton: MatMenuTrigger;
 
   constructor() {}
 
