@@ -12,7 +12,7 @@ module.exports = function(timeStamp) {
         method: 'PUT',
         uri: 'http://127.0.0.1:5984/_users/org.couchdb.user:' + user,
         headers: { 'Content-Type': 'application/json' },
-        body: { name: user, password: 'e2e', roles: [], type: 'user' },
+        body: { name: user, password: 'e2e', roles: [ 'learner' ], type: 'user' },
         json: true
       });
     },
