@@ -14,7 +14,7 @@ exports.config = {
   capabilities: {
     'browserName': 'chrome',
     'chromeOptions': {
-      args: [ "--headless", "--disable-gpu", "--window-size=1280,800" ]
+      args: [ "--headless", "--disable-gpu", "--window-size=1280,800", "--no-sandbox" ]
     }
   },
   directConnect: true,
