@@ -5,8 +5,8 @@ import { PageNotFoundComponent } from './page-not-found/page-not-found.component
 import { AuthService } from './shared/auth-guard.service';
 
 export const routes: Routes = [
-  { path: '', loadChildren: './home/home.module#HomeModule', canActivateChild: [ AuthService ] },
-  { path: 'login', loadChildren: './login/login.module#LoginModule', canActivate: [ AuthService ] },
+  { path: '', loadChildren: () => import('./home/home.module').then(m => m.HomeModule), canActivateChild: [ AuthService ] },
+  { path: 'login', loadChildren: () => import('./login/login.module').then(m => m.LoginModule), canActivate: [ AuthService ] },
   { path: '**', component: PageNotFoundComponent }
 ];
 
