@@ -14,23 +14,29 @@ const routes: Routes = [
   { path: '', component: HomeComponent,
     children: [
       { path: '', component: DashboardComponent },
-      { path: 'users', loadChildren: '../users/users.module#UsersModule' },
-      { path: 'manager', loadChildren: '../manager-dashboard/manager-dashboard.module#ManagerDashboardModule' },
-      { path: 'courses', loadChildren: '../courses/courses.module#CoursesModule' },
-      { path: 'myCourses', loadChildren: '../courses/courses.module#CoursesModule', data: { myCourses: true } },
+      { path: 'users', loadChildren: () => import('../users/users.module').then(m => m.UsersModule) },
+      { path: 'manager', loadChildren: () => import('../manager-dashboard/manager-dashboard.module').then(m => m.ManagerDashboardModule) },
+      { path: 'courses', loadChildren: () => import('../courses/courses.module').then(m => m.CoursesModule) },
+      { path: 'myCourses', loadChildren: () => import('../courses/courses.module').then(m => m.CoursesModule), data: { myCourses: true } },
       { path: 'requests', component: CommunityComponent },
-      { path: 'feedback', loadChildren: '../feedback/feedback.module#FeedbackModule' },
-      { path: 'resources', loadChildren: '../resources/resources.module#ResourcesModule' },
-      { path: 'myLibrary', loadChildren: '../resources/resources.module#ResourcesModule', data: { myLibrary: true } },
-      { path: 'meetups', loadChildren: '../meetups/meetups.module#MeetupsModule' },
+      { path: 'feedback', loadChildren: () => import('../feedback/feedback.module').then(m => m.FeedbackModule) },
+      { path: 'resources', loadChildren: () => import('../resources/resources.module').then(m => m.ResourcesModule) },
+      {
+        path: 'myLibrary',
+        loadChildren: () => import('../resources/resources.module').then(m => m.ResourcesModule), data: { myLibrary: true }
+      },
+      { path: 'meetups', loadChildren: () => import('../meetups/meetups.module').then(m => m.MeetupsModule) },
       { path: 'notifications', component: NotificationsComponent },
-      { path: 'submissions', loadChildren: '../submissions/submissions.module#SubmissionsModule' },
-      { path: 'submissions/:type', loadChildren: '../submissions/submissions.module#SubmissionsModule' },
-      { path: 'mySurveys', loadChildren: '../submissions/submissions.module#SubmissionsModule', data: { mySurveys: true } },
+      { path: 'submissions', loadChildren: () => import('../submissions/submissions.module').then(m => m.SubmissionsModule) },
+      { path: 'submissions/:type', loadChildren: () => import('../submissions/submissions.module').then(m => m.SubmissionsModule) },
+      {
+        path: 'mySurveys',
+        loadChildren: () => import('../submissions/submissions.module').then(m => m.SubmissionsModule), data: { mySurveys: true }
+      },
       { path: 'upgrade', component: UpgradeComponent },
       { path: 'upgrade/myplanet', component: UpgradeComponent, data: { myPlanet: true } },
-      { path: 'teams', loadChildren: '../teams/teams.module#TeamsModule' },
-      { path: 'surveys', loadChildren: '../surveys/surveys.module#SurveysModule' },
+      { path: 'teams', loadChildren: () => import('../teams/teams.module').then(m => m.TeamsModule) },
+      { path: 'surveys', loadChildren: () => import('../surveys/surveys.module').then(m => m.SurveysModule) },
       { path: 'myAchievements', component: UsersAchievementsComponent },
       { path: 'myAchievements/update', component: UsersAchievementsUpdateComponent },
       { path: 'news', component: NewsComponent },
