@@ -13,11 +13,11 @@ const routes: Routes = [
   { path: '', component: ManagerDashboardComponent },
   { path: 'sync', component: ManagerSyncComponent },
   { path: 'fetch', component: ManagerFetchComponent },
-  { path: 'meetups', loadChildren: '../meetups/meetups.module#MeetupsModule', data: { parent: true } },
-  { path: 'courses', loadChildren: '../courses/courses.module#CoursesModule', data: { parent: true } },
-  { path: 'resources', loadChildren: '../resources/resources.module#ResourcesModule', data: { parent: true } },
+  { path: 'meetups', loadChildren: () => import('../meetups/meetups.module').then(m => m.MeetupsModule), data: { parent: true } },
+  { path: 'courses', loadChildren: () => import('../courses/courses.module').then(m => m.CoursesModule), data: { parent: true } },
+  { path: 'resources', loadChildren: () => import('../resources/resources.module').then(m => m.ResourcesModule), data: { parent: true } },
   { path: 'configuration', component: ManagerDashboardConfigurationComponent, data: { update: true } },
-  { path: 'users', loadChildren: '../users/users.module#UsersModule' },
+  { path: 'users', loadChildren: () => import('../users/users.module').then(m => m.UsersModule) },
   { path: 'reports', component: ReportsComponent },
   { path: 'reports/detail', component: ReportsDetailComponent },
   { path: 'reports/pending', component: ReportsPendingComponent },
