@@ -6,7 +6,7 @@ import { TeamsViewComponent } from './teams-view.component';
 const routes: Routes = [
   { path: '', component: TeamsComponent },
   { path: 'view/:teamId', component: TeamsViewComponent },
-  { path: 'users', loadChildren: '../users/users.module#UsersModule' },
+  { path: 'users', loadChildren: () => import('../users/users.module').then(m => m.UsersModule) },
 ];
 @NgModule({
   imports: [ RouterModule.forChild(routes) ],
